pcc: work around uint8_t pcc loop iterator bug

+++ src/mini.c
@@ -58,7 +58,7 @@ void outhex(uint8_t i) {
 }
 
 void outhex32(uint32_t i) {
-   for (uint8_t d = 0; d < 8; d++) {
+   for (int d = 0; d < 8; d++) {
       // Note: this uses ASHC which is a signed shift
       outhex(i >> 28);
       i <<= 4;
